Add Amazon api functionality (Slow), Create Swap and Metadata page

+++ db/schema/products.sql
@@ -6,7 +6,7 @@ CREATE TABLE products (
   number_of_products INT,
   lowest_price INT,
   highest_price INT,
-  avg_star_rating INT,
+  avg_star_rating FLOAT,
   is_prime BOOLEAN,
   user_id INT REFERENCES users(id) ON DELETE CASCADE
   );
